requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.1
pyjwt
passlib[argon2]
python-multipart